refactor implementation into fattn-buffers.cpp
